Extract HoverEffect to a standalone module
