+++ packages/react-native-gesture-handler/src/HoverEffect.ts
@@ -1,0 +1,5 @@
+export enum HoverEffect {
+  NONE = 0,
+  LIFT = 1,
+  HIGHLIGHT = 2,
+}
