+++ packages/react-native-gesture-handler/src/handlers/gestures/hoverGesture.ts
@@ -1,3 +1,4 @@
+import type { HoverEffect } from '../../HoverEffect';
 import type { GestureUpdateEvent } from '../gestureHandlerCommon';
 import type { HoverGestureHandlerEventPayload } from '../GestureHandlerEventPayload';
 import type { BaseGestureConfig } from './gesture';
@@ -7,12 +8,6 @@ export type HoverGestureChangeEventPayload = {
   changeX: number;
   changeY: number;
 };
-
-export enum HoverEffect {
-  NONE = 0,
-  LIFT = 1,
-  HIGHLIGHT = 2,
-}
 
 export interface HoverGestureConfig {
   hoverEffect?: HoverEffect;
