+++ packages/react-native-gesture-handler/src/v3/hooks/gestures/hover/HoverTypes.ts
@@ -1,5 +1,5 @@
 import type { StylusData } from '../../../../handlers/gestureHandlerCommon';
-import type { HoverEffect } from '../../../../handlers/gestures/hoverGesture';
+import type { HoverEffect } from '../../../../HoverEffect';
 import type {
   BaseGestureConfig,
   ExcludeInternalConfigProps,
